Question 1
so I have started with question 1 and am going to go down the line. I
needed some help with this one and found an easyer way to write this
out in my text book.
/^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/;
by using this, I was able to not have to write 16 lines of code. moving
on